+++ gossip/__init__.py
@@ -2,9 +2,10 @@ from .exception_policy import IgnoreExceptions, RaiseDefer, RaiseImmediately
 from .exceptions import NotNowException
 from .groups import (create_group, get_global_group, get_group, get_groups,
                      get_or_create_group, unregister_token)
-from .hooks import (define, get_all_hooks, get_hook, Hook, register, trigger,
+from .hooks import (define, get_all_hooks, get_all_registrations, get_hook, Hook, register, trigger,
                     trigger_with_tags, mute_context
                     )
+from .blueprint import Blueprint
 
 
 def set_exception_policy(policy):